pyproject.toml
[project]
name = "cbx"
version = "1.0.0"
description = "A converter from PDF and CBR to WebP CBZ"
authors = [
    {name = "Fco. Javier Delgado del Hoyo", email = "frandelhoyo@gmail.com"},
]
dependencies = [
    "patool>=2.2.0",
    "Pillow>=9.1.0",
    "tqdm>=4.64.0",
    "pdf2image>=1.16.3",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}


[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project.scripts]
cbx = "cbx.__main__:main"

[tool.pdm.dev-dependencies]
linting = [
    "ruff>=0.1.5",
]
lint = [
    "ruff>=0.2.2",
    "pre-commit>=3.6.1",
]
